pyproject.toml
[tool.poetry]
name = "weni-chats"
version = "0.1.0"
description = ""
authors = ["helllllllder <heldergsds@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
Pillow = "^9.1.1"
django-rest-framework = "^0.1.0"
channels = "^3.0.4"
djongo = "^1.3.6"
django-environ = "^0.8.1"
pymongo = "3.12.1"
black = "^22.3.0"
isort = "^5.10.1"
channels-redis = "^3.4.0"
coreapi = "^2.3.3"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
